api/requirements.txt
yandex_cloud_ml_sdk==0.15.1
pydantic==2.11.9
python-dotenv==1.1.1
